feat: implementar função para remover espaços laterais e quebras de linha do campo Código do formulário de cadastro de produto do Eureka
